feat(ngModelAttrsManipulator): 🎆 Add support for property accessor keys 🎇
This simply replaces the  with the given key and is prefixed with . Unfortunately,
this current approach wont work for braket syntax for the first property, but should cover most use
cases

closes #452 BREAKING CHANGE: We're now altering the ng-model regardless of anything. If you used to
have dots or brackets in your properties, this will break you. Most of you probably wont have a
problem.