feat(formlyConfigProvider): Expose field types map from formlyConfigProvider
For a plugin I am writing that renders an inspector showing the current configuration for formly, I
need access to the types that have been registered with formly. This change exposes the map of type
name to type configuration object via a new method on formlyConfigProvider called getTypes().

This is in support of issue #663