feat(formly-field): Adds parser for keys that contain arrays
angular. does not (and will not https://github.com/angular/angular.js/pull/9850) properly
handle arrays in keys unless they have already been created in the model. This fix/feature adds a
separate parser for these circumstances and a formlyConfig.extras flag to control its use. The flag
is in place to minimize impact on current functionality of the setter.

This is in support of https://github.com/formly-js/angular-formly/issues/706